Clean up and update workspace files
* Remove deprecated TSLint config

* Remove jsconfig.json

jsconfig.json is an alternative name for tsconfig.json that has allowJs
set to true. Since the directory already contains a tsconfig.json file,
this file does nothing.

* Remove Gitpod configuration

Gitpod is not as useful in CE because we're focusing on standalone.
Also, the configuration was added just to promote the service and there
are other similar services which may be preferred by the user.

* Remove Travis CI configuration

Travis CI is not used in shapez anymore in favor of GitHub Actions. In
addition to that, CI setup for the community edition will be done once
the codebase is in a better condition.

* Remove CONTRIBUTORS and CONTRIBUTING.md

The contributors file is out of date and unneeded as GitHub now shows
all collaborators anyway - however it could be still used in the future
to show a list of contributors in the game. CONTRIBUTING.md contains the
text of CLA, which is not enforced in this repository.

* Replace VSC workspaces with regular settings files

VSCode workspaces don't provide much benefit in a source tree like
shapez. This commit also adds EditorConfig and ESLint VSCode extensions
to the recommended extensions list.

* Remove CI workflow for building the game

It hasn't been updated in a while and we don't need it for now. Once the
codebase is in a more managable state, a new workflow will be added.

* Remove some unused resources

There are still many resources left, but they will be removed along with
the relevant code later.